refactor(workspace): split sandbox.rs into per-platform modules
sandbox.rs exceeded the CI line limit (1241 lines). Split into:
- sandbox/mod.rs (629) — shared types, validation, dispatch, cross-platform tests
- sandbox/bwrap.rs (385) — Linux bubblewrap: probe, install hint, prefix builder, tests
- sandbox/seatbelt.rs (189) — macOS Seatbelt: profile builder, tests

No functional changes — pure file reorganization.
